pyproject.toml
[tool.poetry]
name = "django-task-manager"
version = "0.0.1"
description = "Task manager with Django backend"
authors = ["shitcoding <shitcoder@pm.me>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1"
python-dotenv = "^0.20.0"
gunicorn = "^20.1.0"
psycopg2 = { version = "^2.9.3", optional = true }
psycopg2-binary = { version = "^2.9.3", optional = true }
dj-database-url = "^1.0.0"
django-bootstrap4 = "^22.2"
django-filter = "^22.1"
django-crispy-forms = "^1.14.0"
crispy-bootstrap5 = "^0.7"
whitenoise = "^6.2.0"
django-environ = "^0.9.0"

[tool.poetry.extras]
psycopg2 = ["psycopg2"]
psycopg2-binary = ["psycopg2-binary"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
pytest-django = "^4.5.2"
pytest-dotenv = "^0.5.2"
Faker = "^15.1.1"
freezegun = "^1.2.2"
coverage = "^6.5.0"
pytest-sugar = "^0.9.6"
django-debug-toolbar = "^3.8.1"
factory-boy = "^3.2.1"
wemake-python-styleguide = "^0.16.1"
pytest-cov = "^3.0.0"
ipython = "^8.5.0"
django-override-autonow = "^0.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
